수정

+++ php_version/controllers/NotController.php
@@ -34,27 +34,41 @@ class NotController{
         $cnt = $post['not_no'];
 
         if($post['not_no'] == ''){
-            $tmp_sql = "SELECT * FROM `notification`";
-            $temp_result = $this->notTable->get_result($tmp_sql);
-            (string)$cnt = $temp_result->rowCount() + 1;
+            $tmp_sql = "SELECT `not_no` FROM `notification`";
+            $result = $this->notTable->get_result($tmp_sql);
+            foreach ($result as $row) {
+                $tmp = $row['not_no']; 
+            }
+            $cnt = (string)((int)$tmp + 1);
+            echo '$cnt의 타입 ='.gettype($cnt).'<br>';
         }
-
-        if($post['not_det_url'] != ''){
+        
+        if($post['not_det_url'] != $file['not_det_url'] || $file['not_det_url'] != ''){
+            $path = '../img/not/'.$cnt.'/det';
+            if (file_exists($path)){
+                $this->delete_img($path);
+            }
             $post['not_det_url'] = $this->img_uplode($file, 'not_det_url', $cnt);
         }
-        if($post['not_ban_url'] != ''){
+        if($post['not_ban_url'] != $file['not_ban_url'] || $file['not_ban_url'] != ''){
+            $path = '../img/not/'.$cnt.'/ban';
+            if (file_exists($path)){
+                $this->delete_img($path);
+            }
             $post['not_ban_url'] = $this->img_uplode($file, 'not_ban_url', $cnt);
         }
-        if($post['not_pop_url'] != ''){
+        if($post['not_pop_url'] != $file['not_pop_url'] || $file['not_pop_url'] != ''){
+            $path = '../img/not/'.$cnt.'/pop';
+            if (file_exists($path)){
+                $this->delete_img($path);
+            }
             $post['not_pop_url'] = $this->img_uplode($file, 'not_pop_url', $cnt);
         }
-
         return $post;
     }
 
     private function img_uplode(array $file, string $key_name, string $cnt){
         $split = explode("_", $key_name);
-        $filename = $split[1].$cnt;
         $tempFile = $file[$key_name]['tmp_name'];
         $fileTypeExt = explode("/", $file[$key_name]['type']);
         $fileType = $fileTypeExt[0];
@@ -74,7 +88,7 @@ class NotController{
                 exit;
                 break;
         }
-        echo '$filename = '.$filename.'<br>';
+
         if($fileType == 'image'){
             if($extStatus){
                 $split = explode("_", $key_name);
@@ -104,12 +118,13 @@ class NotController{
     }
 
     private function delete_img($path) {
+        echo '$path = '.$path.'<br>';
         $dirs = dir($path);
 
         while(false !== ($entry = $dirs->read())) {
             if(($entry != '.') && ($entry != '..')) {            
                 if(is_dir($path.'/'.$entry)) {
-                    delete_img($path.'/'.$entry);
+                    $this->delete_img($path.'/'.$entry);
                 } else {
                     @unlink($path.'/'.$entry);
                 }
@@ -148,7 +163,9 @@ class NotController{
 
     public function delete(){
         $path = '../img/not/'.$_POST['not_no'];
-        $this->delete_img($path);
+        if (file_exists($path)){
+            $this->delete_img($path);
+        }
 
         $this->notTable->deleteData($_POST['not_no']);
         header('location: /not/list');
@@ -161,13 +178,10 @@ class NotController{
                 $this->notTable->insertData($_POST);
             }
             else{
-                $path = '../img/not/'.$_POST['not_no'];
-                $this->delete_img($path);
-
                 $_POST =  $this->img_manage($_POST, $_FILES);
                 $this->notTable->updateData($_POST);
             }
-            header('location: /not/list');
+            //header('location: /not/list');
         }
         if(isset($_GET['not_no'])){
             $row = $this->notTable->selectID($_GET['not_no']);
